move test classes and complete javadoc
