+++ src/main/java/com/ge/healthcare/dose/services/ruleengine/UnknownRuleSetException.java
@@ -1,5 +1,11 @@
 package com.ge.healthcare.dose.services.ruleengine;
 
+/**
+ * This exception is raised when a non existent RuleSet is find.
+ *
+ * @author Frédéric Delorme<frederic.Delorme@ge.com>
+ * @since 2019
+ */
 public class UnknownRuleSetException extends Exception {
     private String ruleSetName;
