+++ src/test/resources/features/ruleengine-context.feature
@@ -1,0 +1,4 @@
+Feature: Using a RuleContext between multiple Rules
+  Test the RuleContext capabilities.
+
+  Scenario: Define Contextual data element
